vfio: Separate the external mapping implementation
Separate it from the ioctls wrappers.
